+++ packages/stream_video/lib/src/models/call_participant_state.dart
@@ -14,7 +14,7 @@ import 'viewport_visibility.dart';
 class CallParticipantState
     with EquatableMixin
     implements Comparable<CallParticipantState> {
-  const CallParticipantState({
+  CallParticipantState({
     required this.userId,
     required this.roles,
     required this.name,
@@ -27,12 +27,13 @@ class CallParticipantState
     this.connectionQuality = SfuConnectionQuality.unspecified,
     this.isOnline = false,
     this.audioLevel = 0,
+    List<double>? audioLevels,
     this.isSpeaking = false,
     this.isDominantSpeaker = false,
     this.isPinned = false,
     this.reaction,
     this.viewportVisibility = ViewportVisibility.unknown,
-  });
+  }) : audioLevels = audioLevels ?? [audioLevel];
 
   final String userId;
   final List<String> roles;
@@ -45,7 +46,13 @@ class CallParticipantState
   final bool isLocal;
   final SfuConnectionQuality connectionQuality;
   final bool isOnline;
+
+  /// The latest audio level for the user.
   final double audioLevel;
+
+  /// List of the last 10 audio levels.
+  final List<double> audioLevels;
+
   final bool isSpeaking;
   final bool isDominantSpeaker;
   final bool isPinned;
@@ -54,6 +61,8 @@ class CallParticipantState
 
   /// Returns a copy of this [CallParticipantState] with the given fields
   /// replaced with the new values.
+  ///
+  /// If you want to update the audioLevel, consider using [copyWithUpdatedAudioLevels].
   CallParticipantState copyWith({
     String? userId,
     List<String>? roles,
@@ -67,6 +76,7 @@ class CallParticipantState
     SfuConnectionQuality? connectionQuality,
     bool? isOnline,
     double? audioLevel,
+    List<double>? audioLevels,
     bool? isSpeaking,
     bool? isDominantSpeaker,
     bool? isPinned,
@@ -86,11 +96,30 @@ class CallParticipantState
       connectionQuality: connectionQuality ?? this.connectionQuality,
       isOnline: isOnline ?? this.isOnline,
       audioLevel: audioLevel ?? this.audioLevel,
+      audioLevels: audioLevels ?? this.audioLevels,
       isSpeaking: isSpeaking ?? this.isSpeaking,
       isDominantSpeaker: isDominantSpeaker ?? this.isDominantSpeaker,
       isPinned: isPinned ?? this.isPinned,
       reaction: reaction ?? this.reaction,
       viewportVisibility: viewportVisibility ?? this.viewportVisibility,
+    );
+  }
+
+  /// Copies the current state and adds the latest [audioLevel] to the last 10 [audioLevels].
+  CallParticipantState copyWithUpdatedAudioLevels({
+    required double audioLevel,
+    bool? isSpeaking,
+  }) {
+    final levels = audioLevels;
+    levels.add(audioLevel);
+    while (levels.length > 10) {
+      levels.removeAt(0);
+    }
+
+    return copyWith(
+      audioLevel: audioLevel,
+      audioLevels: audioLevels,
+      isSpeaking: isSpeaking,
     );
   }
 
@@ -119,7 +148,7 @@ class CallParticipantState
         'publishedTracks: $publishedTracks, '
         'isLocal: $isLocal, '
         'connectionQuality: $connectionQuality, isOnline: $isOnline, '
-        'audioLevel: $audioLevel, isSpeaking: $isSpeaking, '
+        'audioLevel: $audioLevel, audioLevels: $audioLevels, isSpeaking: $isSpeaking, '
         'isDominantSpeaker: $isDominantSpeaker, isPinned: $isPinned, '
         'reaction: $reaction, viewportVisibility: $viewportVisibility}';
   }
@@ -138,6 +167,7 @@ class CallParticipantState
         connectionQuality,
         isOnline,
         audioLevel,
+        audioLevels,
         isSpeaking,
         isDominantSpeaker,
         isPinned,